cleanup classes
