+++ Src/LibraryCore.Kafka/IKafkaNodeCreator.cs
@@ -1,0 +1,6 @@
+﻿namespace LibraryCore.Kafka;
+
+public interface IKafkaNodeCreator
+{
+    Task CreateNodeAsync(int nodeId, string jobKey, CancellationToken cancellationToken);
+}
